Make a few small optimizations to JRE `ImmutableSet` construction for this "new" world in which we never reuse the array passed to `construct`.
(We entered that world almost 6 years ago in cl/185156965.)

A more significant optimization may be coming in cl/590212390, but it can hurt performance in some cases, so we may or may not ultimately want it.

(Even for this CL, I could imagine bad performance for users who call `ImmutableSet.copyOf(Collections.filter(...))`. If that's an issue in practice, then maybe we should insert a special case for it inside `copyOf`. I at least made sure to call `size()` only once.)

As further testing (which would become even more important with cl/590212390), I've enhanced `testCopyOf_threadSafe` (and enabled it for `Immutable`*`Sorted`*`Set` while I was there).

The benchmarks and original idea are taken from jylee's cl/589183494 and b/315526394.

Note that some of the changes to `ImmutableSetAllocationTest` pass even before this CL. We must have let its numbers get out of date at some point.

RELNOTES=n/a
PiperOrigin-RevId: 592077874